+++ front/src/components/UserEdit.vue
@@ -9,13 +9,8 @@
             修正メニュー
             <v-spacer/>
             <v-btn>
-                <v-icon 
-                    red
-                    @click='aa'
-                >
-                <!-- <router-link to="../user"> -->
+                <v-icon red @click='closeCard'>
                     mdi-close-circle-outline
-                <!-- </router-link> -->
                 </v-icon>
             </v-btn>
         </v-app-bar>
@@ -23,8 +18,12 @@
 
         <v-card-text>
             <v-form ref="loginForm">
+                <v-text-field label="変更後のユーザー名"
+                    prepend-icon="mdi-human"
+                    v-model="model.username" 
+                    :counter="128"/>
 
-                <v-text>
+                <v-text v-if='editmail'>
                     メールアドレス
                 </v-text>
 
@@ -32,19 +31,23 @@
                     prepend-icon="mdi-email"
                     v-model="model.email" 
                     :counter="128"
-                    :rules="emailRules"/>
+                    :rules="emailRules"
+                    v-if='editmail'/>
                 <v-text-field label="変更後のメールアドレス"
                     prepend-icon="mdi-email"
                     v-model="model.email_edit" 
                     :counter="128"
-                    :rules="emailRules2"/>
+                    :rules="emailRules2"
+                    v-if='editmail'/>
 
-                <v-card-actions>
+                <v-card-actions v-if='editmail'>
                     <v-spacer/>
+                    <v-btn v-if="editpassword" @click="changeEmail">変更しない</v-btn>
+                    <v-btn v-if="!editpassword" @click="changeEmail">パスワードを変更</v-btn>
                     <v-btn @click="register">修正</v-btn>
                 </v-card-actions>
                                 
-                <v-text>
+                <v-text v-if='editpassword'>
                     パスワード
                 </v-text>
                 <v-text-field label="現在のパスワードを入力"
@@ -54,7 +57,8 @@
                     @click:append="showPassword = !showPassword"
                     v-model="model.password"
                     :counter="32"
-                    :rules="passwordRules"/>
+                    :rules="passwordRules"
+                    v-if='editpassword'/>
 
                 <v-text-field label="変更後のパスワード"
                     prepend-icon="mdi-lock" 
@@ -63,10 +67,14 @@
                     @click:append="showPassword2 = !showPassword2"
                     v-model="model.password_edit"
                     :counter="32"
-                    :rules="passwordRules2"/>
+                    :rules="passwordRules2"
+                    v-if='editpassword'/>
                 
-                <v-card-actions>
+                <v-card-actions v-if='editpassword'>
                     <v-spacer/>
+                    <v-btn @click="initialState">Clear</v-btn>
+                    <v-btn v-if="editmail" @click="changePassword">変更しない</v-btn>
+                    <v-btn v-if="!editmail" @click="changePassword">メールアドレスを変更</v-btn>
                     <v-btn @click="register">修正</v-btn>
                 </v-card-actions>
             </v-form>
@@ -77,24 +85,15 @@
 
 <script>
 export default {
-
+    props: {
+        user: null
+    },
     data: function() {
         return {
-            user: {
-                name: 'タカタ',
-                user_id: '000000',
-                mail: 'takata@takata.com',
-                password: 'takata',
-                src: require('@/assets/pose_kuyashii_man.png')
-            },
-            model: {
-                username : "",
-                email : "",
-                email_edit: "",
-                password : "",
-                password_edit : "",
-            },
-
+            editer: true,
+            editmail: true,
+            editpassword: true,
+            model: {},
             showPassword : false,
             showPassword2 : false,
             usernameRules: [
@@ -120,13 +119,32 @@ export default {
             ],
         }
     },
-
+    mounted(){
+        this.initialState()
+    },
     methods: {
+        initialState: function() {
+                this.editer= true
+                this.editmail= true
+                this.editpassword= true
+                this.model= {
+                    username : "",
+                    email : "",
+                    email_edit: "",
+                    password : "",
+                    password_edit : "",
+                    edit: false,
+                }
+        },
         register: function() {
             if (this.$refs.loginForm.validate()) {
                 if(this.check_database()) {
                     this.create_account()
-                    this.$router.push('/user')
+                    this.model.edit_email = this.editmail
+                    this.model.edit_password = this.editpassword
+                    this.$emit('close',this.model)
+                    
+                    // this.$router.push('/user')
                 }
                 else {
                     console.log("failed to send database")
@@ -147,11 +165,27 @@ export default {
         create_account: function() {
             //TODO: アカウントを作成する処理
             console.log("create_account")
+            
         },
 
-        aa: function(){
-            this.$router.push('../user')
+        closeCard: function(){
+            console.log('close')
+            
+            this.$emit('close',this.model)
+            // this.$router.push('../user')
+        },
+
+        changeEmail: function(){
+            this.editmail = false
+            this.editpassword = true
+        },
+
+        changePassword: function(){
+            this.editpassword = false
+            this.editmail = true
         }
+    
+
     }
     
 }
